+++ src/Routes/Amenities.js
@@ -10,7 +10,8 @@ const router = Router();
 
 router.get("/", async (req, res, next) => {
   try {
-    const amenities = await getAmenities();
+    const { name } = req.query;
+    const amenities = await getAmenities(name);
     res.json(amenities);
   } catch (error) {
     next(error);
@@ -22,11 +23,13 @@ router.post("/", authMiddleware, async (req, res, next) => {
     const { name } = req.body;
     //If statement tip from stack overflow:
     if (req.body.constructor === Object && Object.keys(req.body).length === 0) {
+      console.log(`Q says: 400 Bad request; Amenity was not created.`);
       res
         .status(400)
         .json({ message: `Q says: 400 Bad request; Amenity was not created.` });
     } else {
       const newAmenity = await createAmenity(name);
+      console.log(`Q says: 201 Created; Amenity created, with name: ${name}`);
       res.status(201).json(newAmenity);
     }
   } catch (error) {
@@ -40,10 +43,12 @@ router.get("/:id", async (req, res, next) => {
     const amenity = await getAmenityById(id);
 
     if (!amenity) {
+      console.log(`Q says: 404 Not found; Amenity with id: ${id}`);
       res
         .status(404)
         .json({ message: `Q says: 404 Not found; Amenity with id: ${id}` });
     } else {
+      console.log(`Q says: 200 Found; Returning amenity with id: ${id}`);
       res.status(200).json(amenity);
     }
   } catch (error) {
@@ -57,11 +62,16 @@ router.delete("/:id", authMiddleware, async (req, res, next) => {
     const deleteAmenity = await deleteAmenityById(id);
 
     if (deleteAmenity) {
+      console.log(`Q says: 200 OK; Amenity with id: ${id} is deleted.`);
       res.status(200).send({
-        message: `Q says: 200 OK; Amenity with id: ${id} is deleted`,
+        message: `Q says: 200 OK; Amenity with id: ${id} is deleted.`,
         deleteAmenity,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Delete failed, for amenity with id: ${id}`
+      );
+
       res.status(404).json({
         message: `Q says: 404 Not found; Delete failed, for amenity with id: ${id}`,
       });
@@ -78,10 +88,14 @@ router.put("/:id", authMiddleware, async (req, res, next) => {
     const updateAmenity = await updateAmenityById(id, { name });
 
     if (updateAmenity) {
+      console.log(`Q says: 200 OK; Amenity with id: ${id} is updated`);
       res.status(200).send({
         message: `Q says: 200 OK; Amenity with id: ${id} is updated`,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Update did not run for, amenity with id: ${id}`
+      );
       res.status(404).json({
         message: `Q says: 404 Not found; Update did not run for, amenity with id: ${id}`,
       });