+++ src/Routes/Bookings.js
@@ -10,7 +10,24 @@ const router = Router();
 
 router.get("/", async (req, res, next) => {
   try {
-    const bookings = await getBookings();
+    const {
+      userId,
+      propertyId,
+      checkinDate,
+      checkoutDate,
+      numberOfGuests,
+      totalPrice,
+      bookingStatus,
+    } = req.query;
+    const bookings = await getBookings(
+      userId,
+      propertyId,
+      checkinDate,
+      checkoutDate,
+      numberOfGuests,
+      totalPrice,
+      bookingStatus
+    );
     res.json(bookings);
   } catch (error) {
     next(error);
@@ -30,6 +47,7 @@ router.post("/", authMiddleware, async (req, res, next) => {
     } = req.body;
     //If statement tip from stack overflow:
     if (req.body.constructor === Object && Object.keys(req.body).length === 0) {
+      console.log(`Q says: 400 Bad request; Booking was not created.`);
       res
         .status(400)
         .json({ message: `Q says: 400 Bad request; Booking was not created.` });
@@ -43,7 +61,18 @@ router.post("/", authMiddleware, async (req, res, next) => {
         totalPrice,
         bookingStatus
       );
-      res.status(201).json(newBooking);
+      if (newBooking) {
+        console.log(
+          `Q says: 201 Created; Booking created, for date: ${checkinDate}`
+        );
+        res.status(201).json(newBooking);
+      } else {
+        //409 as error as there is a conflict with putting data in the database.
+        console.log(`Q says: 409 Conflict; Booking was not created.`);
+        res
+          .status(409)
+          .json({ message: `Q says: 409 Conflict; Booking was not created.` });
+      }
     }
   } catch (error) {
     next(error);
@@ -56,10 +85,12 @@ router.get("/:id", async (req, res, next) => {
     const booking = await getBookingById(id);
 
     if (!booking) {
+      console.log(`Q says: 404 Not found; Booking with id: ${id}`);
       res
         .status(404)
-        .json({ message: `Q says: 404 Not found; Booking with id ${id}` });
+        .json({ message: `Q says: 404 Not found; Booking with id: ${id}` });
     } else {
+      console.log(`Q says: 200 Found; Returning booking with id: ${id}`);
       res.status(200).json(booking);
     }
   } catch (error) {
@@ -73,13 +104,17 @@ router.delete("/:id", authMiddleware, async (req, res, next) => {
     const deleteBooking = await deleteBookingById(id);
 
     if (deleteBooking) {
+      console.log(`Q says: 200 OK; Booking with id: ${id} is deleted`);
       res.status(200).send({
-        message: `Q says: 200 OK; Booking with id ${id} is deleted`,
+        message: `Q says: 200 OK; Booking with id: ${id} is deleted`,
         deleteBooking,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Delete failed, for booking with id: ${id}`
+      );
       res.status(404).json({
-        message: `Q says: 404 Not found; Delete failed, for booking with id ${id}`,
+        message: `Q says: 404 Not found; Delete failed, for booking with id: ${id}`,
       });
     }
   } catch (error) {
@@ -100,7 +135,7 @@ router.put("/:id", authMiddleware, async (req, res, next) => {
       bookingStatus,
     } = req.body;
     if (req.body.constructor === Object && Object.keys(req.body).length === 0) {
-      console.log("Object missing");
+      console.log("Q says: Object missing");
     }
     const updateBooking = await updateBookingById(id, {
       userId,
@@ -113,12 +148,16 @@ router.put("/:id", authMiddleware, async (req, res, next) => {
     });
 
     if (updateBooking) {
+      console.log(`Q says: 200 OK; Booking with id: ${id} is updated`);
       res.status(200).send({
-        message: `Q says: 200 OK; Booking with id ${id} is updated`,
+        message: `Q says: 200 OK; Booking with id: ${id} is updated`,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Update did not run for, booking with id: ${id}`
+      );
       res.status(404).json({
-        message: `Q says: 404 Not found; Update did not run for, booking with id ${id}`,
+        message: `Q says: 404 Not found; Update did not run for, booking with id: ${id}`,
       });
     }
   } catch (error) {