+++ src/Routes/Hosts.js
@@ -11,7 +11,16 @@ const router = Router();
 
 router.get("/", async (req, res, next) => {
   try {
-    const hosts = await getHosts();
+    const { username, name, email, phoneNumber, profilePicture, aboutMe } =
+      req.query;
+    const hosts = await getHosts(
+      username,
+      name,
+      email,
+      phoneNumber,
+      profilePicture,
+      aboutMe
+    );
     res.json(hosts);
   } catch (error) {
     next(error);
@@ -31,9 +40,10 @@ router.post("/", authMiddleware, async (req, res, next) => {
     } = req.body;
     //If statement tip from stack overflow:
     if (req.body.constructor === Object && Object.keys(req.body).length === 0) {
-      res
-        .status(400)
-        .json({ message: `Q says: 400 Bad request; Host was not created.` });
+      console.log(`Q says: 400 Bad request; Host-user was not created.`);
+      res.status(400).json({
+        message: `Q says: 400 Bad request; Host-user was not created.`,
+      });
     } else {
       const newHost = await createNewHost(
         username,
@@ -44,7 +54,16 @@ router.post("/", authMiddleware, async (req, res, next) => {
         profilePicture,
         aboutMe
       );
-      res.status(201).json(newHost);
+      if (newHost) {
+        console.log(`Q says: 201 Created; Host-user created, for: ${name}`);
+        res.status(201).json(newHost);
+      } else {
+        //409 as error as there is a conflict with putting data in the database.
+        console.log(`Q says: 409 Conflict; Host-user was not created.`);
+        res.status(409).json({
+          message: `Q says: 409 Conflict; Host-user was not created.`,
+        });
+      }
     }
   } catch (error) {
     next(error);
@@ -57,10 +76,12 @@ router.get("/:id", async (req, res, next) => {
     const host = await getHostById(id);
 
     if (!host) {
+      console.log(`Q says: 404 Not found; Host-user with id: ${id}`);
       res.status(404).json({
         message: `Q says: 404 Not found; Host-user with id: ${id}`,
       });
     } else {
+      console.log(`Q says: 200 Found; Returning host-user with id: ${id}`);
       res.status(200).json(host);
     }
   } catch (error) {
@@ -74,11 +95,15 @@ router.delete("/:id", authMiddleware, async (req, res, next) => {
     const deleteHost = await deleteHostById(id);
 
     if (deleteHost) {
+      console.log(`Q says: 200 OK; Host-user with id: ${id} is deleted`);
       res.status(200).send({
-        message: `Q says: 200 OK; Host-user with id ${id} is deleted`,
+        message: `Q says: 200 OK; Host-user with id: ${id} is deleted`,
         deleteHost,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Delete failed, for host-user with id: ${id}`
+      );
       res.status(404).json({
         message: `Q says: 404 Not found; Delete failed, for host-user with id: ${id}`,
       });
@@ -111,10 +136,14 @@ router.put("/:id", authMiddleware, async (req, res, next) => {
     });
 
     if (upsateHost) {
+      console.log(`Q says: 200 OK; Host-user with id: ${id} is updated`);
       res.status(200).send({
         message: `Q says: 200 OK; Host-user with id ${id} is updated`,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Update did not run for, host-user with id: ${id}`
+      );
       res.status(404).json({
         message: `Q says: 404 Not found; Update did not run for, host-user with id: ${id}`,
       });