+++ src/Routes/Properties.js
@@ -10,7 +10,28 @@ const router = Router();
 
 router.get("/", async (req, res, next) => {
   try {
-    const properties = await getProperties();
+    const {
+      title,
+      description,
+      location,
+      pricePerNight,
+      bedroomCount,
+      bathRoomCount,
+      maxGuestCount,
+      hostId,
+      rating,
+    } = req.query;
+    const properties = await getProperties(
+      title,
+      description,
+      location,
+      pricePerNight,
+      bedroomCount,
+      bathRoomCount,
+      maxGuestCount,
+      hostId,
+      rating
+    );
     res.json(properties);
   } catch (error) {
     next(error);
@@ -32,6 +53,7 @@ router.post("/", authMiddleware, async (req, res, next) => {
     } = req.body;
     //If statement tip from stack overflow:
     if (req.body.constructor === Object && Object.keys(req.body).length === 0) {
+      console.log(`Q says: 400 Bad request; Property was not created.`);
       res.status(400).json({
         message: `Q says: 400 Bad request; Property was not created.`,
       });
@@ -47,7 +69,18 @@ router.post("/", authMiddleware, async (req, res, next) => {
         hostId,
         rating
       );
-      res.status(201).json(newProperty);
+      if (newProperty) {
+        console.log(
+          `Q says: 201 Created; Property created, with title: ${title}`
+        );
+        res.status(201).json(newProperty);
+      } else {
+        //409 as error as there is a conflict with putting data in the database.
+        console.log(`Q says: 409 Conflict; Property was not created.`);
+        res
+          .status(409)
+          .json({ message: `Q says: 409 Conflict; Property was not created.` });
+      }
     }
   } catch (error) {
     next(error);
@@ -58,12 +91,13 @@ router.get("/:id", async (req, res, next) => {
   try {
     const { id } = req.params;
     const property = await getPropertyById(id);
-
     if (!property) {
+      console.log(`Q says: 404 Not found; Property with id: ${id}`);
       res
         .status(404)
         .json({ message: `Q says: 404 Not found; Property with id: ${id}` });
     } else {
+      console.log(`Q says: 200 Found; Returning property with id: ${id}`);
       res.status(200).json(property);
     }
   } catch (error) {
@@ -75,13 +109,16 @@ router.delete("/:id", authMiddleware, async (req, res, next) => {
   try {
     const { id } = req.params;
     const deleteProperty = await deletePropertyById(id);
-
     if (deleteProperty) {
+      console.log(`Q says: 200 OK; Property with id: ${id} is deleted`);
       res.status(200).send({
         message: `Q says: 200 OK; Property with id: ${id} is deleted`,
         deleteProperty,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Delete failed, for property with id: ${id}`
+      );
       res.status(404).json({
         message: `Q says: 404 Not found; Delete failed, for property with id: ${id}`,
       });
@@ -116,12 +153,15 @@ router.put("/:id", authMiddleware, async (req, res, next) => {
       hostId,
       rating,
     });
-
     if (updateProperty) {
+      console.log(`Q says: 200 OK; Property with id: ${id} is updated`);
       res.status(200).send({
         message: `Q says: 200 OK; Property with id: ${id} is updated`,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Update did not run for, property with id: ${id}`
+      );
       res.status(404).json({
         message: `Q says: 404 Not found; Update did not run for, property with id: ${id}`,
       });