+++ src/Routes/Users.js
@@ -10,7 +10,14 @@ const router = Router();
 
 router.get("/", async (req, res, next) => {
   try {
-    const users = await getUsers();
+    const { username, name, email, phoneNumber, profilePicture } = req.query;
+    const users = await getUsers(
+      username,
+      name,
+      email,
+      phoneNumber,
+      profilePicture
+    );
     res.json(users);
   } catch (error) {
     next(error);
@@ -23,6 +30,7 @@ router.post("/", authMiddleware, async (req, res, next) => {
       req.body;
     //If statement tip from stack overflow:
     if (req.body.constructor === Object && Object.keys(req.body).length === 0) {
+      console.log(`Q says: 400 Bad request; User was not created.`);
       res
         .status(400)
         .json({ message: `Q says: 400 Bad request; User was not created.` });
@@ -36,7 +44,16 @@ router.post("/", authMiddleware, async (req, res, next) => {
         phoneNumber,
         profilePicture
       );
-      res.status(201).json(newUser);
+      if (newUser) {
+        console.log(`Q says: 201 Created; User created, for: ${name}`);
+        res.status(201).json(newUser);
+      } else {
+        //409 as error as there is a conflict with putting data in the database.
+        console.log(`Q says: 409 Conflict; User was not created.`);
+        res
+          .status(409)
+          .json({ message: `Q says: 409 Conflict; User was not created.` });
+      }
     }
   } catch (error) {
     next(error);
@@ -47,12 +64,13 @@ router.get("/:id", async (req, res, next) => {
   try {
     const { id } = req.params;
     const user = await getUserById(id);
-
     if (!user) {
+      console.log(`Q says: 404 Not found; User with id: ${id}`);
       res
         .status(404)
         .json({ message: `Q says: 404 Not found; User with id: ${id}` });
     } else {
+      console.log(`Q says: 200 Found; Returning user with id: ${id}`);
       res.status(200).json(user);
     }
   } catch (error) {
@@ -66,11 +84,15 @@ router.delete("/:id", authMiddleware, async (req, res, next) => {
     const deleteUser = await deleteUserById(id);
 
     if (deleteUser) {
+      console.log(`Q says: 200 OK; User with id: ${id} is deleted`);
       res.status(200).send({
         message: `Q says: 200 OK; User with id: ${id} is deleted`,
         deleteUser,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Delete failed, for user with id: ${id}`
+      );
       res.status(404).json({
         message: `Q says: 404 Not found; Delete failed, for amenity with id: ${id}`,
       });
@@ -95,11 +117,15 @@ router.put("/:id", authMiddleware, async (req, res, next) => {
     });
 
     if (updateUser) {
+      console.log(`Q says: 200 OK; User with id: ${id} is updated`);
       res.status(200).send({
         message: `Q says: 200 OK; User with id: ${id} is updated`,
         updateUser,
       });
     } else {
+      console.log(
+        `Q says: 404 Not found; Update did not run for, user with id: ${id}`
+      );
       res.status(404).json({
         message: `Q says: 404 Not found; Update did not run for, user with id: ${id}`,
       });