+++ src/Services/Bookings/createBookings.js
@@ -12,23 +12,27 @@ const createBooking = async (
   bookingStatus
 ) => {
   const prisma = new PrismaClient();
-  const booking = await prisma.booking.create({
-    data: {
-      userBooking: {
-        connect: { id: userId },
+  //To complete the clean lines while testing (Noticed DB constraints were still possible)
+  try {
+    const booking = await prisma.booking.create({
+      data: {
+        userBooking: {
+          connect: { id: userId },
+        },
+        propertyBooking: {
+          connect: { id: propertyId },
+        },
+        checkinDate,
+        checkoutDate,
+        numberOfGuests,
+        totalPrice,
+        bookingStatus,
       },
-      propertyBooking: {
-        connect: { id: propertyId },
-      },
-      checkinDate,
-      checkoutDate,
-      numberOfGuests,
-      totalPrice,
-      bookingStatus,
-    },
-  });
-
-  return booking;
+    });
+    return booking;
+  } catch (error) {
+    return false;
+  }
 };
 
 export default createBooking;