+++ src/Services/Properties/createProperties.js
@@ -13,23 +13,28 @@ const createProperty = async (
 ) => {
   const prisma = new PrismaClient();
   //With mappings it seems handy to name per column.
-  const property = await prisma.property.create({
-    data: {
-      hostProperty: {
-        connect: { id: hostId },
+  //To complete the clean lines while testing (Noticed DB constraints were still possible)
+  try {
+    const property = await prisma.property.create({
+      data: {
+        hostProperty: {
+          connect: { id: hostId },
+        },
+        title: title,
+        description: description,
+        location: location,
+        pricePerNight: pricePerNight,
+        bedroomCount: bedroomCount,
+        bathRoomCount: bathRoomCount,
+        maxGuestCount: maxGuestCount,
+        rating: rating,
       },
-      title: title,
-      description: description,
-      location: location,
-      pricePerNight: pricePerNight,
-      bedroomCount: bedroomCount,
-      bathRoomCount: bathRoomCount,
-      maxGuestCount: maxGuestCount,
-      rating: rating,
-    },
-  });
+    });
 
-  return property;
+    return property;
+  } catch (error) {
+    return false;
+  }
 };
 
 export default createProperty;