+++ src/Services/Reviews/getReviews.js
@@ -1,8 +1,35 @@
 import { PrismaClient } from "@prisma/client";
 
-const getReviews = async () => {
+const getReviews = async (userId, propertyId, rating, comment) => {
   const prisma = new PrismaClient();
-  const reviews = await prisma.review.findMany();
+  //Revieuws convert ratings and search on minimum rating.
+  //commenths in contian search.
+  rating = rating && parseInt(rating);
+  const recordCount = await prisma.review.count({
+    where: {
+      userId,
+      propertyId,
+      rating: {
+        gte: rating,
+      },
+      comment: {
+        contains: comment,
+      },
+    },
+  });
+  console.log(`Q Says: records found for your host-user query: `, recordCount);
+  const reviews = await prisma.review.findMany({
+    where: {
+      userId,
+      propertyId,
+      rating: {
+        gte: rating,
+      },
+      comment: {
+        contains: comment,
+      },
+    },
+  });
 
   return reviews;
 };