+++ src/Services/Users/createUsers.js
@@ -12,17 +12,22 @@ const createUser = async (
 ) => {
   console.log("user name: ", username);
   const prisma = new PrismaClient();
-  const user = await prisma.user.create({
-    data: {
-      username,
-      password,
-      name,
-      email,
-      phoneNumber,
-      profilePicture,
-    },
-  });
-  return user;
+  //To complete the clean lines while testing (Noticed DB constraints were still possible)
+  try {
+    const user = await prisma.user.create({
+      data: {
+        username,
+        password,
+        name,
+        email,
+        phoneNumber,
+        profilePicture,
+      },
+    });
+    return user;
+  } catch (error) {
+    return false;
+  }
 };
 
 export default createUser;